Se pusieron fotos para las partes donde se hara la reparacion por la cual entro a Status -> Taller

+++ app/src/main/java/com/example/negociomx_hyundai/BE/ParteDanno.kt
@@ -1,7 +1,7 @@
 package com.example.negociomx_hyundai.BE
 
 data class ParteDanno(
-    var IdParteDanno: Int = 0,
+    var IdParteDanno: Short = 0,
     var Nombre: String = "",
     var Activa: Boolean = false,
     var IdTabla: Int? = null,
